+++ chezmoi2/internal/chezmoi/chezmoi_unix.go
@@ -3,15 +3,26 @@
 package chezmoi
 
 import (
+	"bufio"
+	"bytes"
 	"fmt"
+	"net"
 	"os"
 	"path"
 	"path/filepath"
+	"regexp"
 	"strings"
 	"syscall"
+
+	vfs "github.com/twpayne/go-vfs"
 )
 
-var umask os.FileMode
+const fqdnHostnameAddr = "127.0.1.1"
+
+var (
+	umask        os.FileMode
+	whitespaceRx = regexp.MustCompile(`\s+`)
+)
 
 func init() {
 	umask = os.FileMode(syscall.Umask(0))
@@ -28,6 +39,14 @@ func ExpandTilde(path, homeDir string) string {
 	default:
 		return path
 	}
+}
+
+// FQDNHostname returns the FQDN hostname.
+func FQDNHostname(fs vfs.FS) (string, error) {
+	if fqdnHostname, err := etcHostsFQDNHostname(fs); err == nil && fqdnHostname != "" {
+		return fqdnHostname, nil
+	}
+	return lookupAddrFQDNHostname()
 }
 
 // GetUmask returns the umask.
@@ -63,6 +82,27 @@ func TrimDirPrefix(p, dir string) (string, error) {
 	}
 }
 
+// etcHostsFQDNHostname returns the FQDN hostname from parsing /etc/hosts.
+func etcHostsFQDNHostname(fs vfs.FS) (string, error) {
+	etcHostsContents, err := fs.ReadFile("/etc/hosts")
+	if err != nil {
+		return "", err
+	}
+	s := bufio.NewScanner(bytes.NewReader(etcHostsContents))
+	for s.Scan() {
+		text := s.Text()
+		text = strings.TrimSpace(text)
+		if index := strings.IndexByte(text, '#'); index != -1 {
+			text = text[:index]
+		}
+		fields := whitespaceRx.Split(text, -1)
+		if len(fields) >= 2 && fields[0] == fqdnHostnameAddr {
+			return fields[1], nil
+		}
+	}
+	return "", s.Err()
+}
+
 // isExecutable returns if info is executable.
 func isExecutable(info os.FileInfo) bool {
 	return info.Mode().Perm()&0o111 != 0
@@ -71,6 +111,19 @@ func isExecutable(info os.FileInfo) bool {
 // isPrivate returns if info is private.
 func isPrivate(info os.FileInfo) bool {
 	return info.Mode().Perm()&0o77 == 0
+}
+
+// lookupAddrFQDNHostname returns the FQDN hostname by doing a reverse lookup of
+// 127.0.1.1.
+func lookupAddrFQDNHostname() (string, error) {
+	names, err := net.LookupAddr(fqdnHostnameAddr)
+	if err != nil {
+		return "", err
+	}
+	if len(names) == 0 {
+		return "", nil
+	}
+	return strings.TrimSuffix(names[0], "."), nil
 }
 
 // umaskPermEqual returns if two permissions are equal after applying umask.